requirements.txt
streamlit
torch
torchvision
ultralytics
opencv-python
Pillow